requirements.txt
Faker==4.0.3
python-dateutil==2.8.1
PyYAML==5.3.1
six==1.14.0
text-unidecode==1.3
uiautomator==0.3.6
urllib3==1.25.9
